Avoid relative import
There are ways to make this work, but it complicates the story in our
internal libraries, since this package lives at
github.com/Shyp/<ourpackage>/vendor/github.com/dav/go-pivotaltracker,
the relative import of "v5/pivotal" doesn't work as easily.

It's a lot easier (and conventional) to use the full name space.